feat: adhoc code compile [APE-1476] (#1707)

+++ src/ape/managers/compilers.py
@@ -5,6 +5,7 @@ from ethpm_types import ContractType
 from ethpm_types.source import Content
 
 from ape.api import CompilerAPI
+from ape.contracts import ContractContainer
 from ape.exceptions import ApeAttributeError, CompilerError, ContractLogicError
 from ape.logging import logger
 from ape.managers.base import BaseManager
@@ -183,18 +184,60 @@ class CompilerManager(BaseManager):
                         existing_artifact.unlink()
 
                     else:
-                        path = self.project_manager.lookup_path(existing_contract.source_id)
-                        if path and existing_contract.source_id != contract_type.source_id:
-                            error_message = f"{ContractType.__name__} collision '{contract_name}'."
-                            raise CompilerError(error_message)
+                        if existing_contract.source_id:
+                            path = self.project_manager.lookup_path(existing_contract.source_id)
+                            if path and existing_contract.source_id != contract_type.source_id:
+                                error_message = (
+                                    f"{ContractType.__name__} collision '{contract_name}'."
+                                )
+                                raise CompilerError(error_message)
 
-                        elif not path:
-                            # Artifact remaining from deleted contract, can delete.
-                            existing_artifact.unlink()
+                            elif not path:
+                                # Artifact remaining from deleted contract, can delete.
+                                existing_artifact.unlink()
 
                 contract_types_dict[contract_name] = contract_type
 
         return contract_types_dict
+
+    def compile_source(
+        self,
+        compiler_name: str,
+        code: str,
+        settings: Optional[Dict] = None,
+        **kwargs,
+    ) -> ContractContainer:
+        """
+        Compile the given program.
+
+        Usage example::
+
+            code = '[{"name":"foo","type":"fallback", "stateMutability":"nonpayable"}]'
+            contract_type = compilers.compile_source(
+                "ethpm",
+                code,
+                contractName="MyContract",
+            )
+
+        Args:
+            compiler_name (str): The name of the compiler to use.
+            code (str): The source code to compile.
+            settings (Optional[Dict]): Compiler settings.
+            **kwargs: Additional overrides for the ``ethpm_types.ContractType`` model.
+
+        Returns:
+            ``ContractContainer``: A contract container ready to be deployed.
+        """
+        compiler = self.get_compiler(compiler_name, settings=settings)
+        if not compiler:
+            raise ValueError(f"Compiler '{compiler_name}' not found.")
+
+        contract_type = compiler.compile_code(
+            code,
+            base_path=self.project_manager.contracts_folder,
+            **kwargs,
+        )
+        return ContractContainer(contract_type=contract_type)
 
     def get_imports(
         self, contract_filepaths: Sequence[Path], base_path: Optional[Path] = None
